% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torsion.R
\name{torsion}
\alias{torsion}
\title{Torsion Calculation
Calculate the Torsion of a Parametric Curve}
\usage{
torsion(alpha, t)
}
\arguments{
\item{alpha}{A vector of 3 expressions defining the parametric curve}

\item{t}{Evaluation parameter}
}
\value{
Scalar value Torsion of alpha at time t
}
\description{
Torsion Calculation
Calculate the Torsion of a Parametric Curve
}
\details{
alpha must be a vector of 3 parametric expressions with param 't'
}
\examples{
tAlpha <- torsion(c(expression(1 + t + cos(t)*sin(t),1 + 0.1*sin(t) + cos(t) * 0.1,0.1*sin(t))), 2)
}
\author{
Leyton Taylor
}
